validation hooked up to backend
