+++ package.json
@@ -65,7 +65,6 @@
     "date-fns": "^4.1.0",
     "embla-carousel-react": "^8.6.0",
     "input-otp": "^1.4.2",
-    "jsonc-parser": "^3.3.1",
     "lucide-react": "^0.541.0",
     "monaco-editor": "^0.52.2",
     "next": "^15.2.3",
